Avoided infinite loop on same pub/sub topic
Added a way of keeping the node from going into an infinite loop when sending and receiving on the same topic.  I've done this by checking for the same msg.payload that was sent and received. A 1002 ms timer is also involved to clear the lastPayload (what was sent). I'm still not sure about the timer. Also started renaming this from timeout (Pete's original code) to mytimeout (how original) to differentiate the two. More work needs to be done so that this is a proper node-red package.