Got the timer on, off, stop and cancel working
I originally started with Pete Scargill's timeout module. Which was a pretty simple module. But I needed some other features such as setting the countdown and warning times. Such that the timer could be used from teh NR UI. For turning on Smart devices and turning them off at some later time that the user can set from the UI.

One feature that was very important was the ability to reject receiving the same message that the node just sent. This can happen with the same topic is used to subscribe and publish to. This makes it easier for the timer and several devices to use the same topic for control. The other requirements are that the node be able to accept a payload (the device command), a timeout value (in seconds) and a warning value (in seconds). These values will override the defaults set in the node.

At this time the code mostly works though thurough testing has not yet been completed. Some basic unit testing shows the it can handle the simple on, off, stop and cancel commands. And it blocks the resending of the commands just sent out so no endless positive feedback happens.