Update comments, documentation and added 1 more output for seconds remaining
For the rewritten mytimeout.js I've added a new output for seconds remaining on
the timer and it's current state. I've updated the comments to reflect that this
is now less Pete Scargill's code and different from big timer. And I've update
the documentation a bit more to better explain the inputs and outputs of
myteimout. Further work is needed here as the documentation is rather ackward
and could possibly confuse not native speakers.