Major changes, added feature and changed the JSON input timer to timeout.
First the old code was too confusing to fix properly. So I rewrote it. I've
corrected the code to accept "timeout" instead of "timer" in the input JSON.
This is, as per the documentation. I've given TIX (the timer) it's own listener
and I've sort of worked a state engine into the input handler. I still have
some code that handles the selection of the defaults, passed and running
values that needs to move. Second, I've fixed a bunch of problems that have
been reported. I've tested the code with a manual run script. I'm pretty sure
I have the 'hung on warning' problem fixed in this release.