Updated the documentation and added a note on a work around
I'm currently learning about branches and trying to figure out how they work. SO I'm getting a bit confused. But I've found that I had a lot of updates to the documentation. Probably to clarify a few things and I've added a work around that user: tsprivate found when running 12 timers in the same flow with the same name. His work around is to move each timer to it's own flow and give each a unique name. I'm not sure if the work around is the A, B or both but recommending both for now. I will be testing this shortly. It's just tought to setup a test framework that handles 12 timers in some random fashion.