Updated with debug option, v2.1.7
Hmm, seems I have a bit to learn on dealing with git branches and merges.

I've fixed the debug option handling. It worked if the node config set the debug value but the default setting is undefined (until you define it - hint). So, I set a default vlaue (0) and also updated the node.status (using this.status) to stopped: init.

I was unable to merge in later bug-timer-overload changes after I merge the first set of changes. I'll need to do some more reading on branches and merges.