Modified newMsg to handle msg.payload { payload, timeout, warning } properly
A user is sending the actual msg object in the msg.payload. So I
modified the newMsg() function to properly merge the msg.payload
object into the msg object. I also needed to modify the set of the
code to return the original msg (with updates to the payload, of
course). At the moment the stop() function has a bit of a kludge using
the line variable like a context variable. This will require a rewrite
of parts of the node. I haven't figured out how I want to do this yet.