Added new feature to ignore case on input match.
Changed the debug into a checkbox instead of text box. Unfortunately  this change in the HTML file does cause some issues with the deploying of changes but it is a message you can ignore. Also, After fixing the msg output and correcting the handling of the timeout and warning, in the last release, I noticed that it was possible that the user might send the output back to the input of the timer but change the case. I'm not sure why they'd need to do this but ignoring the case match is simple enough to add. And with that I bumped this to v2.2.0