Feature request: adjusted code to allow a blank safe msg to not be sent.
obbers has requested a feature request to  allow the "Timer On Payload" to work
like the "Warning state payload". Where is the default safe is blank, no safe
message is sent.

I change the code so safe message no longer defaults to "on" and when blank
it no longer sends an on message. The timer starts as it normally does and
the off message is sent when the timer counts down to 0. Warning is still
optional.