Updated the n1 flow settings to include the timeout and warning
I've now updated the 'on' tests to include the outsafe: "on", outwarning: "warning", outunsafe: "off", warning: "5", and timer: "30". I'm not sure why the defaults are not set when run as a test. I'll need to add a new node to a flow and see if it works without being set.