Added new pause feature and passed the existing test coverage
In 2018 Alan Reader asked that I add a pause feature to mytimeout. Alan created a pull and I let it sit. So now a few releases later I'm found I needed to manually add the pause feature. I think I have it added now and I've renamed the 'tick' variable 'tixID' which will avoid confusion with the 'ticks' variable and the 'tixID' name better reflects it's ture meaning.

So far the initial add has passed the various tests I've created. I now need to add test coverage for the pause feature. I'd also like to better define the requirements for this feature as they are pretty loose at the moment.