I've added a Mocha test suite for partial testing coverage.
I've added a Mocha test suite for partial testing coverage. I need to work on the requirements and a grid to make sure I'm testing the correct requirements. At the moment this is a work in progress and has a lot of junk tests. Since I'm also learning how to test a node-red flow some parts may be wrong. Some of the tests now cover the new pause feature and I'm very confused by Alan's original requirements. The pause/suspnd code and where it gets runs seem out of line with a count down timer.